added field restriction to inputs

+++ db/migrate/20120712002626_create_cart_products.rb
@@ -2,7 +2,7 @@ class CreateCartProducts < ActiveRecord::Migration
   def change
     create_table :cart_products do |t|
       t.integer :cart_id
-      t.decimal :quantity
+      t.integer :quantity
       t.integer :product_id
       t.integer :bind_id
       t.string :options
